+++ src/fastertransformer/utils/ShapeCheck.cc
@@ -20,7 +20,7 @@ bool CheckShapeConsistent(const std::vector<Shape>& shape_list) {
 std::string ShapeStringView(const Shape& shape) {
     std::string s;
     s = s + '(';
-    for (int i = 0; i < shape.size(); i++) {
+    for (size_t i = 0; i < shape.size(); i++) {
         s =  s + std::to_string(shape[i]) + ',';
     }
     s = s + ')';
